requirements.txt
# =============================================
# blog-automation — Python 패키지 의존성
# ohopick + ahapick 공통
# =============================================

# Gemini AI (텍스트 생성)
google-genai>=1.70.0

# Google Blogger API (OAuth 인증 + 포스팅)
google-api-python-client>=2.193.0
google-auth>=2.49.1
google-auth-oauthlib>=1.3.1
google-auth-httplib2>=0.3.1

# 이미지 처리 + Cloudinary 업로드
Pillow>=10.0.0
cloudinary>=1.44.1

# HTTP 요청 (이미지 다운로드, RSS 수집)
requests>=2.33.0

# 한국어 번역 (오호픽: 퍼머링크·이미지 검색어 생성)
deep-translator>=1.11.4

# HTML 파싱
beautifulsoup4>=4.14.0
lxml>=4.9.0
